[update] api request function to separate file
